+++ FacturacionAPI/Controllers/LineasFacturaController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Security.Claims;
 using FacturacionAPI.DTOs.Facturas;
 using FacturacionAPI.DTOs.Lineas;
+using FacturacionAPI.Extensions;
 using inercya.EntityLite;
 using Microsoft.AspNetCore.Mvc;
 using Proyecto_Facturas.Data;
@@ -8,7 +9,7 @@ using Proyecto_Facturas.Data;
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace FacturacionAPI.Controllers
-{ 
+{
     [Route("api/[controller]")]
     [ApiController]
     public class LineasFacturaController : ControllerBase
@@ -43,6 +44,12 @@ namespace FacturacionAPI.Controllers
                     return BadRequest("Los datos de la linea son nulos.");
                 }
                 int userId = User.GetUserId();
+
+                if (!validateInvoiceUser(nuevaLinea.IdFactura, userId))
+                {
+                    return StatusCode(403, $"El usuario no tiene acceso a esa factura");
+                }
+
                 var fechaActual = DateTime.UtcNow;
                 nuevaLinea.Creado = fechaActual;
                 nuevaLinea.Modificado = fechaActual;
@@ -51,16 +58,22 @@ namespace FacturacionAPI.Controllers
 
                 await _dataService.LineaFacturaRepository.SaveAsync(nuevaLinea);
                 var facturaResumen = await ActualizarFacturaAsync(nuevaLinea.IdFactura, userId, fechaActual);
-                var lineaCompleta =  await _dataService.LineaFacturaRepository
+                var lineaCompleta = await _dataService.LineaFacturaRepository
                 .GetAsync(LineaFacturaProjections.BaseTable, nuevaLinea.IdLineaFactura);
 
                 if (_dataService.IsActiveTransaction)
                     _dataService.Commit();
 
-                return Ok(new LineaFacturaResponseDto{
+                return Ok(new LineaFacturaResponseDto
+                {
                     Linea = lineaCompleta,
                     Factura = facturaResumen,
-                        });
+                });
+            }
+            catch (InvalidOperationException ex)
+            {
+                if (this._dataService.IsActiveTransaction) this._dataService.Rollback();
+                return StatusCode(415, $"La factura se ha modificado en otra sesion, es necesario recargar: {ex.Message}");
             }
             catch (Exception ex)
             {
@@ -79,9 +92,16 @@ namespace FacturacionAPI.Controllers
                 {
                     return BadRequest("Los datos de la linea son nulos.");
                 }
+                int userId = User.GetUserId();
+
+                if (!validateInvoiceUser(nuevaLinea.IdFactura, userId))
+                {
+                    return StatusCode(403, $"El usuario no tiene acceso a esa factura");
+                }
+
                 var lineaOriginal = await _dataService.LineaFacturaRepository
                 .GetAsync(LineaFacturaProjections.BaseTable, nuevaLinea.IdLineaFactura);
-                int userId = User.GetUserId();
+
                 var fechaActual = DateTime.UtcNow;
                 lineaOriginal.Modificado = fechaActual;
                 lineaOriginal.ModificadoPor = userId;
@@ -101,9 +121,14 @@ namespace FacturacionAPI.Controllers
                     Factura = facturaResumen,
                 });
             }
+            catch (InvalidOperationException ex)
+            {
+                if (this._dataService.IsActiveTransaction) this._dataService.Rollback();
+                return StatusCode(415, $"La factura se ha modificado en otra sesion, es necesario recargar: {ex.Message}");
+            }
             catch (Exception ex)
             {
-                if(this._dataService.IsActiveTransaction) this._dataService.Rollback();
+                if (this._dataService.IsActiveTransaction) this._dataService.Rollback();
                 return StatusCode(500, $"Error al guardar la factura: {ex.Message}");
             }
         }
@@ -114,17 +139,26 @@ namespace FacturacionAPI.Controllers
             _dataService.BeginTransaction();
             try
             {
-
                 int userId = User.GetUserId();
                 var fechaActual = DateTime.UtcNow;
-                var lineaActual =  await _dataService.LineaFacturaRepository.GetAsync(LineaFacturaProjections.BaseTable, id);
+                var lineaActual = await _dataService.LineaFacturaRepository.GetAsync(LineaFacturaProjections.BaseTable, id);
                 var idFactura = lineaActual.IdFactura;
+
+                if (!validateInvoiceUser(idFactura, userId))
+                {
+                    return StatusCode(403, $"El usuario no tiene acceso a esa factura");
+                }
 
                 await _dataService.LineaFacturaRepository.DeleteAsync(id);
                 var facturaResumen = await ActualizarFacturaAsync(idFactura, userId, fechaActual);
                 if (_dataService.IsActiveTransaction)
                     _dataService.Commit();
                 return Ok(facturaResumen);
+            }
+            catch (InvalidOperationException ex)
+            {
+                if (this._dataService.IsActiveTransaction) this._dataService.Rollback();
+                return StatusCode(415, $"La factura se ha modificado en otra sesion, es necesario recargar: {ex.Message}");
             }
             catch (Exception ex)
             {
@@ -139,6 +173,10 @@ namespace FacturacionAPI.Controllers
                 .GetAsync(FacturaProjections.BaseTable, idFactura);
 
             if (factura == null) return null;
+
+            if (!validateInvoiceStatus(factura)) {
+                throw new InvalidOperationException();
+            }
 
             var lineas = await _dataService.LineaFacturaRepository
                 .Query(LineaFacturaProjections.BaseTable)
@@ -166,6 +204,24 @@ namespace FacturacionAPI.Controllers
             };
         }
 
-    }
+        private Boolean validateInvoiceUser(int facturaId, int userId)
+        {
 
+            return _dataService.FacturaRepository
+                      .Query(FacturaProjections.BaseTable)
+                      .Where(FacturaFields.IdFactura, facturaId)
+                      .And(FacturaFields.CreadoPor, userId)
+                      .Any();
+
+        }
+
+        private Boolean validateInvoiceStatus(Factura factura)
+        {
+            if (factura.Status != InvoiceStatusExtension.statusToId(Enums.InvoiceStatusEn.EnCreacion))
+            {
+                return false;
+            }
+            return true;
+        }
+    }
 }
